test 0.1 ready to build

+++ src/sysid_pkg/src/SysIDTools.h
@@ -11,20 +11,21 @@
 using namespace std;
 
 /**
- * @short Read input data from CSV and place in map object
+ * @short Read input data from CSV and place in a map
+ * @details First column of CSV must be integers represting miliseconds
  */
 map load_data(string filepath) {
 	//
 	// create hashmap object
 	//
-	map<float,vector<float>> m;
+	map<int,vector<float>> m;
 	//
 	// read data from CSV file
 	//
 	ifstream indata;
 	indata.open(path);
 	string line;
-	float t = 0.0f;
+	int t = 0;
 	while (getline(indata, line)) {
 		vector<float> values;
 		stringstream lineStream(line);
@@ -33,7 +34,7 @@ map load_data(string filepath) {
 		// get the first "cell" as the time
 		//
 		getline(lineStream, cell, ',');
-		t = stof(cell);
+		t = stoi(cell);
 		//
 		// the rest of the line is the vector of values
 		while (getline(lineStream, cell, ',')) {
